+++ geonode/resource/settings.py
@@ -19,4 +19,4 @@
 #########################################################################
 import os
 
-RESOURCE_MANAGER_CONCRETE_CLASS = os.environ.get('RESOURCE_MANAGER_CONCRETE_CLASS', 'geonode.resource.geoserver.GeoServerResourceManager')
+RESOURCE_MANAGER_CONCRETE_CLASS = os.environ.get('RESOURCE_MANAGER_CONCRETE_CLASS', 'geonode.geoserver.manager.GeoServerResourceManager')
